+++ fili-generic-example/src/main/java/com/yahoo/wiki/webservice/data/config/dimension/GenericDimensionConfigs.java
@@ -17,6 +17,8 @@ import java.util.Collections;
 import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Set;
+import java.util.Map;
+import java.util.HashMap;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
 
@@ -24,7 +26,7 @@ import java.util.stream.Collectors;
  * Hold all the dimension configurations for a generic druid configuration.
  */
 public class GenericDimensionConfigs {
-    private final Set<DimensionConfig> dimensionConfigs;
+    private final Map<String, Set<DimensionConfig>> dataSourceToDimensionConfigs;
 
     /**
      * Construct the dimension configurations.
@@ -32,24 +34,28 @@ public class GenericDimensionConfigs {
      * @param configLoader  Supplies DataSourceConfigurations to build the dimensions from.
      */
     public GenericDimensionConfigs(Supplier<List<? extends DataSourceConfiguration>> configLoader) {
-        dimensionConfigs = configLoader.get().stream()
-                .flatMap(tableName -> tableName.getDimensions().stream())
-                .map(dimensionName -> new DefaultKeyValueStoreDimensionConfig(
-                        () -> dimensionName,
-                        dimensionName,
-                        "",
-                        dimensionName,
-                        "General",
-                        getDefaultFields(),
-                        getDefaultKeyValueStore(dimensionName),
-                        getDefaultSearchProvider(dimensionName)
-                ))
-                .collect(
-                        Collectors.collectingAndThen(
-                                Collectors.toSet(),
-                                Collections::unmodifiableSet
-                        )
-                );
+        dataSourceToDimensionConfigs = new HashMap<>();
+        configLoader.get()
+                .forEach(dataSourceConfiguration -> {
+                    Set<DimensionConfig> tableDimensionConfigs = dataSourceConfiguration.getDimensions().stream()
+                            .map(dimensionName -> new DefaultKeyValueStoreDimensionConfig(
+                                            () -> dimensionName,
+                                            dimensionName,
+                                            "",
+                                            dimensionName,
+                                            "General",
+                                            getDefaultFields(),
+                                            getDefaultKeyValueStore(dimensionName),
+                                            getDefaultSearchProvider(dimensionName)
+                                    )
+                            ).collect(
+                                    Collectors.collectingAndThen(
+                                            Collectors.toSet(),
+                                            Collections::unmodifiableSet
+                                    ));
+
+                    dataSourceToDimensionConfigs.put(dataSourceConfiguration.getName(), tableDimensionConfigs);
+                });
     }
 
     /**
@@ -58,13 +64,27 @@ public class GenericDimensionConfigs {
      * @return set of dimension configurations
      */
     public Set<DimensionConfig> getAllDimensionConfigurations() {
-        return dimensionConfigs;
+        return dataSourceToDimensionConfigs.values().stream()
+                .flatMap(Set::stream)
+                .distinct()
+                .collect(Collectors.toSet());
+    }
+
+    /**
+     * Get all the dimension configurations associated with this datasource.
+     *
+     * @param dataSourceConfiguration  The datasource configuration's dimensions to load
+     *
+     * @return the dimension configurations for this datasource
+     */
+    public Set<DimensionConfig> getDimensionConfigs(DataSourceConfiguration dataSourceConfiguration) {
+        return dataSourceToDimensionConfigs.getOrDefault(dataSourceConfiguration.getName(), Collections.emptySet());
     }
 
     /**
      * Lazily provide a KeyValueStore for this store name.
      *
-     * @param storeName  the name for the key value store
+     * @param storeName  The name for the key value store
      *
      * @return A KeyValueStore instance
      */